chore(refactor): further moving out of bits of model to entities folder
